requirements_ssh.txt
paramiko
pyinstaller
cryptography
pynput
pywin32
psutil
python-xlib
uvicorn